# KAFKA-14336: MetadataResponse#convertToNodeArray uses iteration