--- a/clients/src/main/java/org/apache/kafka/common/requests/MetadataResponse.java
+++ b/clients/src/main/java/org/apache/kafka/common/requests/MetadataResponse.java
@@ -177,8 +177,10 @@ public class MetadataResponse extends AbstractResponse {
     }
 
     private static Node[] convertToNodeArray(List<Integer> replicaIds, Map<Integer, Node> nodesById) {
-        Node[] nodes = new Node[replicaIds.size()];
-        for (int i = 0; i < replicaIds.size(); i++) {
+        // Since this is on hot path for partition info, use indexed iteration to avoid allocation overhead of Streams.
+        int size = replicaIds.size();
+        Node[] nodes = new Node[size];
+        for (int i = 0; i < size; i++) {
             Integer replicaId = replicaIds.get(i);
             Node node = nodesById.get(replicaId);
             if (node == null)