--- a/jmh-benchmarks/src/main/java/org/apache/kafka/jmh/common/MetadataResponseBenchmark.java
+++ b/jmh-benchmarks/src/main/java/org/apache/kafka/jmh/common/MetadataResponseBenchmark.java
@@ -53,6 +53,7 @@ import java.util.stream.IntStream;
 @BenchmarkMode(Mode.AverageTime)
 @OutputTimeUnit(TimeUnit.NANOSECONDS)
 public class MetadataResponseBenchmark {
+
     @Param({"10", "500", "1000"})
     private int nodes;
 
@@ -66,8 +67,8 @@ public class MetadataResponseBenchmark {
                 Optional.of(4),
                 Optional.of(42),
                 IntStream.range(0, nodes).boxed().collect(Collectors.toList()),
-                IntStream.range(0, nodes).filter(i1 -> i1 % 3!=0).boxed().collect(Collectors.toList()),
-                IntStream.range(0, nodes).filter(i2 -> i2 % 3==0).boxed().collect(Collectors.toList()));
+                IntStream.range(0, nodes).filter(i1 -> i1 % 3 != 0).boxed().collect(Collectors.toList()),
+                IntStream.range(0, nodes).filter(i2 -> i2 % 3 == 0).boxed().collect(Collectors.toList()));
         nodesById = new HashMap<>(nodes);
         for (int i = 0; i < nodes; i++) {
             nodesById.put(i, new Node(i, "localhost", 1234));