@schlosna 2024-08-14
Updated to address formatting checkstyle flagged