@kirktrue 2024-08-14
FWIW, here are my results from testing this change locally:

VM version: `JDK 21.0.4, OpenJDK 64-Bit Server VM, 21.0.4+7-LTS`
Computer: `2023 MacBookPro M3 Pro`

Before optimization:

```
Benchmark                                  (nodes)  Mode  Cnt     Score    Error  Units
MetadataResponseBenchmark.toPartitionInfo       10  avgt    5   134.025 ±  3.085  ns/op
MetadataResponseBenchmark.toPartitionInfo      500  avgt    5  3615.080 ± 30.166  ns/op
MetadataResponseBenchmark.toPartitionInfo     1000  avgt    5  9216.628 ± 43.761  ns/op
```

With optimization:

```
Benchmark                                  (nodes)  Mode  Cnt     Score    Error  Units
MetadataResponseBenchmark.toPartitionInfo       10  avgt    5   103.968 ±  3.418  ns/op
MetadataResponseBenchmark.toPartitionInfo      500  avgt    5  2796.510 ± 35.638  ns/op
MetadataResponseBenchmark.toPartitionInfo     1000  avgt    5  7059.945 ± 82.241  ns/op
```

I'm seeing a pretty consistent speedup of ~22-23% 🎉