@schlosna 2024-08-14
Thanks for the reviews everyone. Incorporated comment & extracted `size()`.